+++ dev/src/dev.clj
@@ -31,4 +31,3 @@
 (integrant.repl/set-prep! #(duct/prep-config (read-config) profiles))
 
 (timbre/set-level! :debug)
-(timbre/info "set-level! :debug")