+++ src/typing_ex/typing.cljs
@@ -9,22 +9,25 @@
    [reagent.core :refer [atom]]
    [reagent.dom :as rdom]))
 
-(def ^:private version "1.1.0")
+(def ^:private version "1.2.0")
 
 (defonce app-state (atom {:text "wait a little"
                           :answer ""
                           :counter 60
                           :errors 0}))
-
 (defonce how-many-typing (atom 0))
+(defonce first-key (atom false))
+
+;; こういうのにはコメントしとかないと。
+;; report-alert 回数練習したら一度、アラートを出す。
+;; この場所で定義するのがいいのか？
 (def ^:private report-alert 10)
 
 (defn reset-app-state! []
   (go (let [response (<! (http/get (str "/drill/" 0)))]
         (swap! app-state assoc :text (:body response))))
-  (swap! app-state assoc :answer "")
-  (swap! app-state assoc :counter 60)
-  (swap! app-state assoc :errors 0))
+  (swap! app-state assoc :answer "" :counter 60 :errors 0)
+  (reset! first-key false))
 
 ;; 0.5.3-SNAPSHOT
 ;; + counter if finished.
@@ -46,48 +49,41 @@
   (let [s1 (str users_nick " さんのスコアは " pt " 点です。")
         s2 (condp < pt
              100 "すばらしい。最高点取れた？平均で 80 点越えよう。"
-             90 "がんばった。もう少しで 100 点だね。"
-             60 "だいぶ上手です。この調子でがんばれ。"
-             30 "指先を見ずに、ゆっくり、ミスを少なく。"
-             "練習あるのみ。")]
+              90 "がんばった。もう少しで 100 点だね。"
+              60 "だいぶ上手です。この調子でがんばれ。"
+              30 "指先を見ずに、ゆっくり、ミスを少なく。"
+              "練習あるのみ。")]
     (str s1 "\n" s2)))
 
+;; it worked!
 (defn send-score []
-  (if (zero? (count (:answer @app-state)))
-    (do
-      (js/alert "チャレンジ、忘れてるよ。打つべし！")
-      (reset-app-state!))
-    ;; (go (let))で書いた部分は並列性を持つ、これが非同期ってことか、
-    ;; 周りと同じように書いたら混乱する。
-    ;; ロジック、実行の順番に注意してプログラムすること。
-    (go (let [response
-              ;; FIXME: CSRF
-              ;; 本来は post だが、CSRF 問題がクリアできず、
-              ;; get で実装している。
-              ;; request ヘッダ中に見つかる
-              ;; ring.middleware.anti-forgery/anti-forgery-token を
-              ;; 利用できないか？
-              (<! (http/get (str "/score?pt=" (pt @app-state))))]
-          (reset-app-state!)
-          (js/alert (nick-pt-message (read-string (:body response))))
-          (swap! how-many-typing inc)
-          (when (= 0 (mod @how-many-typing report-alert))
-            (js/alert "がんばってんねー。一旦、休憩入れたら？"))))))
+  (go (let [token (.-value (js/document.getElementById "__anti-forgery-token"))
+            response (<! (http/post
+                          "/score"
+                          {:form-params
+                            {:pt (pt @app-state)
+                             :__anti-forgery-token token}}))]
+        (js/alert (nick-pt-message (read-string (:body response))))
+        (reset-app-state!))))
+  ;; (swap! how-many-typing inc)
+  ;; (when (= 0 (mod @how-many-typing report-alert))
+  ;;   (js/alert "がんばってんねー。一旦、休憩入れたら？")))
 
 (defn count-down []
-  (swap! app-state update :counter dec)
-  (when (neg? (:counter @app-state))
-    (swap! app-state update :counter inc)
-    (send-score)))
-
-(js/setInterval count-down 1000)
+  (when @first-key
+    (swap! app-state update :counter dec)
+    (when (neg? (:counter @app-state))
+      (swap! app-state update :counter constantly 0)
+      (send-score))))
 
 (defn by-dots [n]
   (take n (repeat "🥶"))) ;;🙅💧💦💔❌🦠🥶🥺
 
 (defn check-key [key]
-  (when (= key "Backspace")
-    (swap! app-state update :errors inc)))
+  (if (= key "Backspace")
+    (swap! app-state update :errors inc)
+    (when-not @first-key
+      (swap! first-key not))))
 
 (defn error-component []
   [:p (by-dots (:errors @app-state))])
@@ -110,22 +106,25 @@
    [error-component]
    [:div
     [:input {:type  "button"
-             :id "counter"
+             :id    "counter"
              :class "btn btn-primary btn-sm"
              :value (:counter @app-state)
              :on-click send-score}] " 🔚全部打ち終わってクリックするとボーナス"]
-   [:ul
-    [:li [:a {:href "/scores"} "scores"]]
-    [:li [:a {:href "/nickname"} "nickname"]
-     "（変更すると過去データが消える）"]
-    [:li [:a {:href "/password"} "password"]
-     "（忘れるとログインできない）"]]
-   [:p[:a {:href "/logout" :class "btn btn-warning btn-sm"} "logout"]]
+  ;;  [:ul
+  ;;   [:li [:a {:href "/nickname"} "nickname"]
+  ;;    "（変更すると過去データが消える）"]
+  ;;   [:li [:a {:href "/password"} "password"]
+  ;;    "（忘れるとログインできない）"]]
+   [:p
+    [:a {:href "/scores" :class "btn btn-primary btn-sm"} "scores"]
+    " "
+    [:a {:href "/logout" :class "btn btn-warning btn-sm"} "logout"]]
    [:hr]
    [:div "hkimura, " version]])
 
 (defn start []
-  (rdom/render [ex-page] (js/document.getElementById "app")))
+  (rdom/render [ex-page] (js/document.getElementById "app"))
+  (js/setInterval count-down 1000))
 
 (defn ^:export init []
   (reset-app-state!)
@@ -145,4 +144,5 @@
 (defn stop []
   ;; stop is called before any code is reloaded
   ;; this is controlled by :before-load in the config
-  (js/console.log "stop"))
+  (js/console.log "stop")
+  (js/setInterval count-down 9999999))