+++ src/typing_ex/view/page.clj
@@ -6,9 +6,7 @@
    [ring.util.anti-forgery :refer [anti-forgery-field]]
    [taoensso.timbre :as timbre :refer [debug]]))
 
-(def ^:private version "1.1.0")
-
-(timbre/set-level! :debug)
+(def ^:private version "1.2.0")
 
 (defn page [& contents]
  [::response/ok
@@ -123,6 +121,7 @@
 ;; see handler.core/scores
 ;; 7days, 30days must sync with the code.
 (defn scores-page [ret nick days]
+  (timbre/debug "scores-page")
   (page
    [:h2 "Typing: Scores(last " days " days)"]
    [:p
@@ -159,7 +158,7 @@
                            (= users_nick nick) "yes"
                            (= users_nick "hkimura") "hkimura"
                            :else "other")}
-                 users_nick]])])
+              users_nick]])])
    [:p
     [:a {:href "/" :class "btn btn-primary btn-sm"} "Go!"]
     " "
@@ -205,7 +204,7 @@
 (defn active-users-page [ret]
   (page
    [:h2 "Typing: active users"]
-   (debug "active-users-pate" ret)
+   ;;(debug "active-users-page" ret)
    (into [:ol]
          (for [[u & _] ret]
            [:li (:users_nick u) " " (ss (:timestamp u))]))))